+++ src/index.html
@@ -33,25 +33,25 @@
       <section class="products">
           <ul class="picons">
             <li>
-              <a href="/src/product-listing/index.html?category=tents">
+              <a href="/product-listing/index.html?category=tents">
                 <img src="https://byui-cit.github.io/wdd330/images/category-icons/noun_Tent_3544767.svg" alt="Tent Icon from Noun Project: Mustofa Bayu" width="100%">
                 <h2>Tents</h2>
               </a>
             </li>
             <li>
-              <a href="/src/product-listing/index.html?category=backpacks">
+              <a href="/product-listing/index.html?category=backpacks">
                 <img src="https://byui-cit.github.io/wdd330/images/category-icons/noun_carrier%20bag_3544771.svg" alt="Sleeping Bag Icon from Noun Project: Mustofa Bayu" width="100%">
                 <h2>Backpacks</h2>
               </a>
             </li>
             <li>
-              <a href="/src/product-listing/index.html?category=sleeping-bags">
+              <a href="/product-listing/index.html?category=sleeping-bags">
                 <img src="https://byui-cit.github.io/wdd330/images/category-icons/noun_Sleeping%20Bag_3544775.svg" alt="Backpack Icon from Noun Project: Mustofa Bayu" width="100%">
                 <h2>Sleeping Bags</h2>
               </a>
             </li>
             <li>          
-              <a href="/src/product-listing/index.html?category=hammocks">
+              <a href="/product-listing/index.html?category=hammocks">
                 <img 
                 src="https://byui-cit.github.io/wdd330/images/category-icons/noun_Hammock_791143.svg" 
                 alt="Hammock Icon from Noun Project: Paul Richard" width="100%">
@@ -66,6 +66,6 @@
       </section>
     </main>
     <footer id="main-footer"></footer>
-    <script src="js/main.js" type="module"></script>
+    <!-- <script src="js/main.js" type="module"></script> -->
   </body>
 </html>
