+++ src/js/ProductList.mjs
@@ -1,42 +1,36 @@
 import { renderListWithTemplate } from "./utils.mjs";
 
-function productCardTemplate(product){
-    return `<li class="product-card">
-    <a href="../product_pages/index.html?product=${product.Id}">
-    <img src="${product.Images.PrimaryMedium}" alt="Image of ${product.Name}">
-    <h3 class="card__brand">${product.Brand.Name}</h3>
-    <h2 class="card__name">${product.Name}</h2>
-    <p class="product-card__price">$${product.FinalPrice}</p></a>
-    </li>`;
+function productCardTemplate(product) {
+  return `<li class="product-card">
+  <a href="/product_pages/index.html?product=${product.Id}">
+  <img
+    src="${product.Images.PrimaryMedium}"
+    alt="Image of ${product.Name}"
+  />
+  <h3 class="card__brand">${product.Brand.Name}</h3>
+  <h2 class="card__name">${product.Name}</h2>
+  <p class="product-card__price">$${product.FinalPrice}</p></a>
+</li>`;
 }
 
-// function filterList(list, count) {
-//     return list.slice(0,count)
-// }
-
-export default class ProductListing {
-    constructor(category, dataSource, listElement){
-        this.category = category;
-        this.dataSource = dataSource;
-        this.listElement = listElement;
-
-    }
-    async init() {
-        const list = await this.dataSource.getData(this.category);
-        // const filteredList = filterList(list, 4);
-        this.renderList(list)
-        document.querySelector(".title").innerHTML = this.category;
-        // console.log(this.category)
-        
-    }
-
-    renderList(list) {
-        renderListWithTemplate(productCardTemplate, this.listElement, list);
-    }
-    async init() {
-        const list = await this.dataSource.getData(this.category);
-        this.renderList(list);
-      }
+export default class ProductList {
+  constructor(category, dataSource, listElement) {
+    // We passed in this information to make our class as reusable as possible.
+    // Being able to define these things when we use the class will make it very flexible
+    this.category = category;
+    this.dataSource = dataSource;
+    this.listElement = listElement;
+  }
+  async init() {
+    // our dataSource will return a Promise...so we can use await to resolve it.
+    const list = await this.dataSource.getData(this.category);
+    // render the list
+    this.renderList(list);
+    //set the title to the current category
+    document.querySelector(".title").innerHTML = this.category;
+  }
+  // render after doing the first stretch
+  renderList(list) {
+    renderListWithTemplate(productCardTemplate, this.listElement, list);
+  }
 }
-
-
