+++ src/js/product-listing.js
@@ -1,0 +1,15 @@
+import ProductData from "./ProductData.mjs";
+import ProductListing from "./ProductList.mjs";
+import { loadHeaderFooter, getParam } from "./utils.mjs";
+
+loadHeaderFooter();
+
+const category = getParam("category");
+const dataSource = new ProductData();
+const title = category.charAt(0).toUpperCase() + category.slice(1);
+document.querySelector(".title").textContent = title;
+const listElement = document.querySelector('.product-list');
+// then create an instance of our ProductList class and send it the correct information.
+const myList = new ProductListing(category, dataSource, listElement);
+// finally call the init method to show our products
+myList.init();
