+++ src/product-listing/index.html
@@ -5,6 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Sleep Outside | Product Listing</title>
     <link rel="stylesheet" href="/src/css/style.css" />
+    <script type="module" src="/product-listing.js"></script>
   </head>
   <body>
     <header class="divider" id="main-header"></header>
@@ -15,6 +16,5 @@
       </section>
     </main>
     <footer id="main-footer"></footer>
-    <script src="../js/product-listing.mjs" type="module"></script>
   </body>
 </html>
